refactor: re-structured submission form to avoid over-rendering
Since the introduction of upload progress, the profile submission form
would re-render all its inner components each time the `progress` event
was dispatched. This caused flickering on the `Video` component.

Memoizing `Video` would fix the symptom, but not the root cause, which
was that the `SubmitProfileCard` component was too big.
