+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - refactor: changed withdraw submission button ([3dbaa53](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/3dbaa53))
 - refactor: extracted withdraw submission button into a component ([ba6f3a5](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/ba6f3a5))
 - refactor: made scripts responsive to the environment ([86d2b65](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/86d2b65))
+- refactor: re-structured submission form to avoid over-rendering ([4eeb2d3](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/4eeb2d3))
 - refactor: removed webcam button from video and photo upload (#104) ([1352cb5](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/1352cb5)), closes [#104](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/issues/104)
 - refactor(next-router): pick network from env var as a default for safety ([cb7b64b](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/cb7b64b))
 - refactor(submission-card): change tampered data notice ([a9f63b1](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/a9f63b1))
@@ -14,6 +15,7 @@
 - refactor(ubi-card): change UBI Accrued to UBI ([cb47c49](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/cb47c49))
 - style: fixed Chrome black outline for buttons on top navbar ([f44df36](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/f44df36))
 - style: fixed connect button background color ([9650e5c](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/9650e5c))
+- style: fixed name and address spacing ([e1a37b3](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/e1a37b3))
 - style: made submit button width fixed ([0093a85](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/0093a85))
 - style: remove comma ([d9aab49](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/d9aab49))
 - fix: add webcam back without mirroring ([1085e47](https://github.com/Proof-Of-Humanity/proof-of-humanity-web/commit/1085e47))
