+++ components/image.js
@@ -3,11 +3,9 @@ import ReactLoadingSkeleton from "react-loading-skeleton";
 import { Image as _Image, useThemeUI } from "theme-ui";
 
 const Image = forwardRef((props, ref) => {
-  const {
-    theme: {
-      images: { [props.variant]: { width, height } = {} },
-    },
-  } = useThemeUI();
+  const { theme } = useThemeUI();
+  const { width, height } = theme.images?.[props.variant] ?? {};
+
   return props.src ? (
     <_Image ref={ref} {...props} />
   ) : (
@@ -19,5 +17,7 @@ const Image = forwardRef((props, ref) => {
     />
   );
 });
+
 Image.displayName = "Image";
+
 export default Image;
